fix(Whole Project): renamed files.
Renamed all files and fixed imports according to Dart Preferences.

fixes #160
