+++ lib/about.dart
@@ -1,5 +1,5 @@
-import 'package:doclense/MainDrawer.dart';
-import 'package:doclense/Providers/ThemeProvider.dart';
+import 'package:doclense/main_drawer.dart';
+import 'package:doclense/Providers/theme_provider.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
