+++ lib/filter_view.dart
@@ -1,14 +1,14 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:doclense/Providers/ThemeProvider.dart';
+import 'package:doclense/Providers/theme_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:photofilters/photofilters.dart';
 import 'package:path/path.dart';
 import 'package:image/image.dart' as image_lib;
 import 'package:provider/provider.dart';
-import 'Providers/ImageList.dart';
+import 'Providers/image_list.dart';
 import 'multi_select_delete.dart';
 
 class FilterImage extends StatefulWidget {
