+++ lib/main_drawer.dart
@@ -1,4 +1,4 @@
-import 'package:doclense/StarredDocuments.dart';
+import 'package:doclense/starred_documents.dart';
 import 'package:doclense/settings.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
@@ -7,9 +7,9 @@ import 'package:url_launcher/url_launcher.dart';
 import 'package:rating_dialog/rating_dialog.dart';
 import 'package:provider/provider.dart';
 
-import 'About.dart';
+import 'about.dart';
 
-import 'Providers/ThemeProvider.dart';
+import 'Providers/theme_provider.dart';
 
 class MainDrawer extends StatelessWidget {
   @override
