+++ lib/settings.dart
@@ -1,10 +1,10 @@
-// import 'package:doclense/Constants/ThemeConstants.dart';
+// import 'package:doclense/Constants/theme_constants.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'MainDrawer.dart';
-import 'Providers/ThemeProvider.dart';
+import 'main_drawer.dart';
+import 'Providers/theme_provider.dart';
 import 'setting_text.dart';
 
 class SettingsScreen extends StatefulWidget {
